+++ rest_framework_pydantic_ai/__init__.py
@@ -2,6 +2,7 @@
 
 from rest_framework_pydantic_ai.spec_toolset import SpecToolset
 from rest_framework_pydantic_ai.types.agent_deps import AgentDeps
+from rest_framework_pydantic_ai.types.query_param import QueryParam
 from rest_framework_pydantic_ai.version import __version__
 
-__all__ = ["AgentDeps", "SpecToolset", "__version__"]
+__all__ = ["AgentDeps", "QueryParam", "SpecToolset", "__version__"]